Refactor Supabase client import paths; move client creation to ssr directory
